Updated track, dueling melodies, guitar play data, closes #55
New track with the same name. 80_4_synth1.ogg and 80_4_synth2.ogg
are the melodies that overlay 80_4.ogg. 80_4.mid contains the
play data in MIDI format. Note values start at 0 and are in groups
of 5. There are 2 instruments and 3 difficulties each, so 30 note
values are utilized total (0-29). Synth 1 is 0-14. Synth 2 is 15-29.
Difficulties increase from easy to medium to hard. So, Synth 1 easy
is 0-4.